add an option that display indexer version
