+++ Editor/Core/Integration.cs
@@ -1,0 +1,13 @@
+﻿using UnityEditor;
+using UnityEngine;
+
+namespace RV
+{
+	[InitializeOnLoad]
+	internal class Integration
+	{
+		static Integration()
+		{
+		}
+	}
+}
