+++ Editor/Core/RefData.cs
@@ -8,39 +8,13 @@ namespace RV
 {
 	internal class RefData
 	{
-		internal struct Version
-		{
-			// 010000
-			public uint major;
-			// 0100
-			public uint minor;
-			// 01
-			public uint maintenance;
-
-			public Version(uint version)
-			{
-				this.major = version / 10000;
-				this.minor = version % 10000 / 100;
-				this.maintenance = version % 100;
-			}
-
-			public static implicit operator uint(Version version)
-			{
-				return version.major * 10000 + version.minor * 100 + version.maintenance;
-			}
-
-			public static implicit operator Version(uint version)
-			{
-				return new Version(version);
-			}
-		}
-		
 		public string guid;
 		
 		public List<string> ownGuids = new List<string>();
 		public List<string> referedByGuids = new List<string>();
 
-		private Version version;
+		public Version version { get; protected set; }
+
 
 		public RefData(string guid)
 		{
@@ -52,6 +26,11 @@ namespace RV
 			string path = FileSystem.CacheDirectory + $"/{guid}.ref";
 			
 			BinaryWriter w = new BinaryWriter(new FileStream(path, FileMode.Create, FileAccess.Write));
+			
+			// version for integration
+			version = ReferenceSetting.INDEX_VERSION;
+			w.Write(version);
+			
 			int ownCount = ownGuids.Count;
 			w.Write(ownCount);
 					
@@ -67,9 +46,6 @@ namespace RV
 			{
 				w.Write(referedByGuids[i]);
 			}
-
-			// 맨 앞에 인덱싱 해야하는데... 방법을 고민
-			w.Write(version);
 					
 			w.Close();
 		}
@@ -94,6 +70,8 @@ namespace RV
 
 			BinaryReader r = new BinaryReader(File.OpenRead(path));
 			
+			asset.version = r.ReadUInt32();
+			
 			int ownCount = r.ReadInt32();
 			for (int i = 0; i < ownCount; i++)
 			{
@@ -104,18 +82,6 @@ namespace RV
 			for (int i = 0; i < byCount; i++)
 			{
 				asset.referedByGuids.Add(r.ReadString());
-			}
-
-			bool isEndOfStream = r.BaseStream.Position == r.BaseStream.Length;
-			if (isEndOfStream)
-			{
-				// 0.0.x 버전
-				asset.version = 0;
-			}
-			else
-			{
-				// 버전 파싱
-				asset.version = r.ReadUInt32();
 			}
 			
 			r.Close();
